feat: implement modular home page components with optimized animations and styling
